split the polyfile tests into separate files
